requirements.txt
flask
flask[async]
loguru
websockets
python-osc
pyyaml
pyserial
